kta/requirements.txt
flask
gunicorn
gevent
python-consul